Changes to frontend & backend
Modified the backend logic to get a random page of movies so that you don't get the same movie choices every time you visit the page.

Modified the frontend code such that you can skip choices or select a movie (max 5 selected) which then will give you an overview of the categories.