Created part of the algoritm
Created one part of the Algoritm logic. This part will take all of the counter-elements of the users choices and create a % based system where it will fill a list of 15 movies and then give a movie a % of those 15 movies.

Ex if Action has a 20% stake in all of the points, then action should take up around 15% of the spots in the 15 item list that is returned.